+++ man-roxygen/Cohorts.R
@@ -1,4 +1,0 @@
-#' @param cohorts A data frame describing the cohorts used in the study. Should
-#'                include the target, comparator, and outcome of interest cohorts.
-#'                The data frame should at least have a `cohortId` and `cohortName`
-#'                columns.